+++ Guest.h
@@ -4,6 +4,8 @@
 #include "Ticket.h"
 #include "Person.h"
 #include "TicketMaster.h"
+#include "fileHelper.h"
+#include "General.h"
 #include <stdlib.h>
 #include <stdio.h>
 #include <crtdbg.h> // TODO: remove before release
@@ -16,7 +18,7 @@ typedef struct _Guest{
 	Ticket* ticket;
 } Guest;
 
-Person* initGuest(char* name, float height, int age); // constructor
+Person* initGuest(char* name, double height, int age); // constructor
 void freeGuest(Person* guest); // destructor
 Person* initGuestByUser(TicketMaster* tMaster);
 void printGuest(const Person* guest);
